+++ src/main/java/net/ryanland/colossus/command/regular/CommandBuilder.java
@@ -13,7 +13,7 @@ public @interface CommandBuilder {
 
     String description();
 
-    int cooldown() default 2;
+    float cooldown() default 2;
 
     boolean guildOnly() default true;
